# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Mandara API
  version: 0.1.0
  description: |
    Mandara is a self-hosted light-novel reading server.

    This document is the single source of truth for the HTTP API contract:
    the backend implements it and the frontend generates its types from it
    (see `frontend/package.json`, script `api-types`).

    Model overview: `books` are pure metadata entries; `book_files` are the
    actual books (uploads or virtual plugin books). Several files can share
    one metadata entry. Chapters, reading sessions and shares attach to
    files. Every file is either `private` (owner + admins only) or `public`
    (all logged-in users).
servers:
  - url: http://127.0.0.1:8080
    description: Local development server (default)
...
paths:
...
  /api/plugins/instances:
    post:
      tags: [plugins]
      summary: |
        Register a plugin instance (admin only). `wasm_file` must be a
        basename returned by GET /api/plugins/wasm-files; `config` is
        validated against the plugin's config schema (defaults fill the
        rest). An explicit `id` becomes the source id used in
        `book_files.source` — stable ids let metadata plugins point their
        `content-source` at a content plugin's instance. A declare-capable
        instance is synced right away.
      operationId: registerInstance
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [wasm_file]
              properties:
                id:
                  type: string
                  description: 'Explicit instance id (default: generated uuid)'
                wasm_file:
                  type: string
                config:
                  type: object
                  description: Config values keyed by schema field keys
      responses:
        "201":
          description: The registered instance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PluginInstance"
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/files/{id}:
    get:
      tags: [files]
      summary: File detail with chapter titles and its book metadata
      operationId: getFile
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "200":
          description: File detail (plugin chapter titles materialized lazily)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileDetail"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [files]
      summary: Toggle visibility or rename the edition label (owner or admin)
      operationId: patchFile
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FileId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchFile"
      responses:
        "200":
          description: Updated file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileMeta"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [files]
      summary: Delete a file (owner or admin)
      operationId: deleteFile
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "204":
          description: Deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
...
    FileId:
      name: id
      in: path
      required: true
      description: Book file id
      schema:
        type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ValidationError:
      description: Structured config validation errors (field-by-field for inline form display)
      content:
        application/json:
          schema:
            type: object
            required: [error, errors]
            properties:
              error:
                type: string
              errors:
                type: array
                items:
                  $ref: "#/components/schemas/ConfigError"
...
    Forbidden:
      description: Not allowed to perform this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource does not exist (or share expired)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Visibility:
      type: string
      enum: [private, public]
    BookMeta:
      type: object
      required: [id, title, authors, description, cover_url, ext, created_by, created_at, series_id, volume_no]
      properties:
        id:
          type: string
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        description:
          type: [string, "null"]
        cover_url:
          type: [string, "null"]
        ext:
          $ref: "#/components/schemas/BookExt"
        created_by:
          description: Id of the user who created this metadata entry
          type: [string, "null"]
        created_at:
          type: string
          format: date-time
        series_id:
          description: |
            Series this book is a volume of (null = standalone). Volume
            splits assign every 卷 its own book entry plus one series
            row; manual assignment (`PATCH /api/books/{id}`) moves books
            in and out of series.
          type: [string, "null"]
        volume_no:
          description: 1-based volume ordinal inside the series (0 = standalone)
          type: integer
          minimum: 0
    BookExt:
      type: object
      description: |
        Extended book metadata (bangumi/douban-style). Every known key is
        optional; `null`/absent = unset. Unknown keys are preserved
        verbatim so future server fields round-trip through old clients.
        `isbn` is stored canonically as a hyphenless ISBN-13 (valid
        ISBN-10 inputs are converted); invalid ISBNs are rejected with
        400.
      properties:
        subtitle:
          description: 副标题
          type: [string, "null"]
        original_title:
          description: 原作名
          type: [string, "null"]
        isbn:
          description: Canonical hyphenless ISBN-13
          type: [string, "null"]
        publisher:
          description: 出版社
          type: [string, "null"]
        pub_date:
          description: 出版时间 (stored exactly as supplied)
          type: [string, "null"]
        translators:
          description: 译者
          type: [array, "null"]
          items:
            type: string
        illustrators:
          description: 插画师
          type: [array, "null"]
          items:
            type: string
        pages:
          description: 页数
          type: [integer, "null"]
          minimum: 0
        price:
          description: 定价 (currency as printed)
          type: [string, "null"]
        binding:
          description: 装帧 (文库/单行本/…)
          type: [string, "null"]
        language:
          description: BCP-47 language hint ("zh-CN", "ja")
          type: [string, "null"]
      additionalProperties: true
...
    FileMeta:
      type: object
      required:
        - id
        - book_id
        - source
        - external_id
        - content_source
        - content_external_id
        - format
        - label
        - visibility
        - owner_id
        - chapter_count
        - created_at
        - volume_no
        - volume_offset
      properties:
        id:
          type: string
        book_id:
          type: string
        source:
          description: "'local' for uploads, otherwise the plugin id"
          type: string
        external_id:
          description: Id within the source (same as `id` for local files)
          type: string
        content_source:
          description: |
            Instance id whose chapters back this file when different from
            `source` (metadata/content source separation); null = `source`.
          type: string
          nullable: true
        content_external_id:
          description: |
            Book id inside the content source; null = `external_id`.
          type: string
          nullable: true
        format:
          type: string
          enum: [epub, txt, plugin]
        label:
          description: Short human label for this edition
          type: string
        visibility:
          $ref: "#/components/schemas/Visibility"
        owner_id:
          type: [string, "null"]
        chapter_count:
          type: integer
        created_at:
          type: string
          format: date-time
        volume_no:
          description: |
            Volume slice of the source book this file covers (0 = the
            whole book). A multi-volume plugin book is acquired as one
            file per 卷.
          type: integer
          minimum: 0
        volume_offset:
          description: |
            Flat chapter index of this volume's first chapter inside the
            source (persisted because the plugin is stateless); lazy
            chapter pulls use `volume_offset + idx`.
          type: integer
          minimum: 0
        original:
          description: |
            Retained original file bytes (`data/files/`, served by
            `GET /api/files/{id}/download`). null = plugin chapter-mode
            files and uploads made before original retention.
          type: object
          nullable: true
          required: [ext, size, sha256]
          properties:
            ext:
              description: File extension without the dot (epub / txt)
              type: string
            size:
              type: integer
            sha256:
              type: string
    ChapterMeta:
      type: object
      required: [idx, title]
      properties:
        idx:
          type: integer
        title:
          type: string
    TocNode:
      type: object
      description: >-
        One node of the hierarchical table of contents. `idx` is the
        chapter index this entry points to (`null` for pure group nodes);
        leaf entries reference entries of the `chapters` list.
      required: [title, idx, children]
      properties:
        title:
          type: string
        idx:
          type: [integer, "null"]
        frag:
          type: string
          nullable: true
          description: >
            Id fragment of the source href (`file.html#section`) when this
            entry points into the middle of its chapter; the reader
            scrolls to the element with that id.
        children:
          type: array
          items:
            $ref: "#/components/schemas/TocNode"
...
    FileDetail:
      type: object
      required: [file, book, chapters, toc]
      properties:
        file:
          $ref: "#/components/schemas/FileMeta"
        book:
          $ref: "#/components/schemas/BookMeta"
        chapters:
          type: array
          items:
            $ref: "#/components/schemas/ChapterMeta"
        toc:
          type: array
          items:
            $ref: "#/components/schemas/TocNode"
...
    PatchFile:
      type: object
      properties:
        visibility:
          $ref: "#/components/schemas/Visibility"
        label:
          type: string
...
    PluginInstance:
      description: One registered plugin instance (source id = `id`)
      type: object
      required: [id, wasm_file, name, enabled, capabilities, source_info]
      properties:
        id:
          description: Source id used in `book_files.source`
          type: string
        wasm_file:
          description: Basename of the wasm component backing this instance
          type: string
        name:
          description: Human-readable plugin name (from the plugin itself)
          type: string
        enabled:
          type: boolean
        capabilities:
          description: >-
            Subset of [declare, search, lookup, identify, content, book-file].
            book-file instances materialize whole book files through the
            upload parser (file mode wins over chapter mode).
          type: array
          items:
            type: string
        source_info:
          description: |
            Interaction hints declared by the plugin (`source-info` export):
            how the frontend should present this source when acquiring
            books. `kind` = "search" | "manual-id" | "browse" | ""
            (empty = derive from capabilities).
          type: object
          required: [kind, id_kind]
          properties:
            kind:
              type: string
            id_kind:
              type: string
            id_hint:
              type: string
              nullable: true
            search_hint:
              type: string
              nullable: true
...
    ConfigError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
...